requirements.txt
transformers>=4.44.0
accelerate>=0.33.0
datasets>=2.20.0
peft>=0.12.0
bitsandbytes>=0.43.0; platform_system != "Windows"
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
langchain>=0.2.3
langchain-community>=0.2.3
langchain-text-splitters>=0.2.1
pydantic>=2.7.0
fastapi>=0.111.0
uvicorn>=0.30.0
typer>=0.12.0
pyyaml>=6.0.1
tqdm>=4.66.4
rich>=13.7.1